+++ plugins/main.lua
@@ -1,4 +1,4 @@
--- RoRR Modding Toolkit v1.0.1
+-- RoRR Modding Toolkit v1.0.8
 
 log.info("Successfully loaded ".._ENV["!guid"]..".")
 
@@ -7,11 +7,13 @@ RoRR_Modding_Toolkit = true
 require("./actor")
 require("./buff")
 require("./callback")
+require("./equipment")
 require("./helper")
 require("./initialize")
 require("./instance")
 require("./item")
 require("./net")
+require("./object")
 require("./player")
 require("./resources")
 require("./survivor")
@@ -20,8 +22,12 @@ require("./alarm")
 -- Testing
 --require("./testing")
 
+
 -- ========== Initialize ==========
 
 function __initialize()
+    Actor.__initialize()
+    Buff.__initialize()
     Instance.__initialize()
+    Item.__initialize()
 end